Move CorrelationDropdown to molecules folder
